Move scalar function from commons to plugins package
