+++ pinot-common/src/main/java/org/apache/pinot/common/function/scalar/JsonFunctions.java
@@ -20,7 +20,7 @@ package org.apache.pinot.common.function.scalar;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import java.util.Map;
-import org.apache.pinot.common.function.annotations.ScalarFunction;
+import org.apache.pinot.spi.annotations.ScalarFunction;
 import org.apache.pinot.spi.utils.JsonUtils;
