+++ pinot-spi/src/main/java/org/apache/pinot/spi/annotations/ScalarFunction.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.pinot.common.function.annotations;
+package org.apache.pinot.spi.annotations;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
